IMU gyro auto zero fix.

+++ node.h
@@ -121,7 +121,7 @@ typedef struct {
 }__attribute__((packed)) _capture;
 //-----------------------------------------------------------------------------
 typedef uint8_t   _ft_option;
-typedef uint32_t  _ft_uint;
+typedef uint16_t  _ft_uint;
 typedef float     _ft_float;
 typedef _ctr      _ft_ctr;
 typedef _pwm      _ft_pwm;
